- Changed A_ChangeFlag so that it doesn't need to alter the flag   string. With strings being passed as names this is unsafe. - Removed unused parameter types from the function parameter parser for   DECORATE. - Changed: All actor name parameters in DECORATE are now passed as   FNames, not as strings. - Fixed: The MAPINFO parser stored the RedirectType as a type pointer.   But at this point DECORATE hasn't been read yet so this was limited to   the internal classes. - Fixed: TXT_NEED_IDCARD wasn't terminated with a ';'. - Fixed: Strife's DeadRebel was missing its DoomEdNum. - With names as type identifiers it is no longer necessary to remap   the monster types to internal constants in A_BossDeath. - Fixed: A_BossDeath got the string from a name - just to get a name from   the string. Using the name directly is sufficient.
SVN r103 (trunk)

+++ src/info.h
@@ -86,7 +86,7 @@ const BYTE SF_BIGTIC	 = 0x80;
 // The first 2 parameters for each function call represent 
 // the old misc1/misc2 values, even for non-weapons
 
-extern TArray<intptr_t> StateParameters;
+extern TArray<int> StateParameters;
 
 struct FState
 {
